app/requirements.txt
flask==2.3.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation-flask==0.43b0
requests==2.31.0 